Add GET test results and add report page output
A GET handler was added for retrieving test results with URL of the
form: /v1/results/{test_id}.
The report page can now be used to view results. It should be accessed
with URL of the form: /output.html?test_id={test_id}. This is just to
keep it simple for now.

Change-Id: I28e9ebf74cc7a6dad360232840b49c6aae27dab0