Added file of required tests to pass to testr
This commit adds a 2015.03 directory to hold artifacts related
to capabilities and tests as approved by the board. The directory
is seeded with a file that containes a sorted list of required
tests from the 2015.03 capabilities. This file is meant to be
passed to Tempest in a testr run to test only the required
defcore capabilities.

Change-Id: Ie51b2256d528d2cd2c7ea2dbbfe98123df27817c