Clarify JSON vs RST authority
During the initial drafting of the 2015A guidelines, we had a
healthy discussion of whether the text (RST) or JSON version
of the capabilities guidelines should be authoritative, and
what the roles of the DefCore Committee and Board of Directors
was intended to be.[1][2]  Further iterations on the matter
were deferred to a seperate patch in order to establish a base
version of the 2015A process to iterate on before it moves from
"Draft" status to "Approved" status.

This patch adjusts the wording of a couple of sections of the
2015A spec in accordance to what I beleive was the outcome of
those discussions.  The patch clarifies that the Board approves
an intended set of capabilities listed in the RST file, and the
DefCore Committee is then tasked with managing the definitions
of those capabilities in the JSON file (with some limitations to
prevent dramatic changes since the Review period) without the need
for Board re-approval.  The patch also clarifies that after approval,
the Committee has the ability to flag tests (presumably in response
to vendor grievances) but can't add additional tests within the same
guideline.

[1] https://review.openstack.org/#/c/165216/4/process/2015A.rst
[2] https://etherpad.openstack.org/p/DefCoreScale.9

Change-Id: I5381099fcb505ca6fb92f96f3a5d61c810f6a569