Flag tests that require SSH
Currently tests that require SSH are not gated against in Tempest
due to the variety of conflicting network models available to
OpenStack clusters. Tests that require SSH need to be rationalized
and refactored, and is a current work item for the openstack-qa
team.

Until this test infrastructure is refactored, the tests should be
flagged but kept for future capability testing.

Change-Id: I9e79090f70648661ab236f67be7bbcd99bc0afc3