Added Designated Sections (copied from Refstack site)
As per DefCore discussion, I've moved the DS JSON over
to the guidelines.

Unlike the original version, I've added the concept
of required, advisory, depricated, removed and
informational so that it matches the process flows.

To be complete, I've done it to all three Guidelines
since they were all the same.

Change-Id: Ia5a906df5228bf34bb24c96e730e2452743a17c8