Introduce "TC-approved release" in process
The OpenStack Foundation bylaws mandate the following:

The Technical Committee shall designate a subset of the OpenStack
Project an “OpenStack Technical Committee Approved Release” from
time to time. The Board of Directors may determine "Trademark
Designated OpenStack Software" from time to time, which will be a
subset of the "OpenStack Technical Committee Approved Release".

Therefore it is necessary that capabilities and designated sections
used in trademark guidelines are chosen from a superset of projects
code and features designated by the TC and called "the TC-approved
Release".

This change proposes to introduce the "TC-approved release" in the
Defcore process to make sure we don't miss that critical step.

Change-Id: Ie99c6ae7f9da590469a22155ccb286d789cc0ebf