Generate RST Guidelines from JSON source
As per the process, we want to use generated RST
instead of hand created files.

In generating the files, I found a comma error in 2015.03
and removed spaces from 2015.04 so there are minor
changes to those files.

Because 2015.next is still in process, it does not make
sense to generate the RST file at this time.  We do
NOT want to have to maintain multiple files for each
change.  I've removed the 2015.next.rst file.

Also added check to make sure we're using the
correct JSON source using the hardcoded DefCore
repo path as an extra check.

Change-Id: I443f4c1cb84eb2c42ad9e6914cee0ce6d498a9bb