Update description of designated section updates.
This update to the process for designated sections removes explicit
mention of participation by OpenStack technical leadership.  I feel
that the only explicit participation by technical leadership in this
whole process should be the TC defining the TC-approved release, as
mandated by the bylaws.  All other participation should be considered
optional and under general community participation and feedback, at
the discretion of individuals.

The existing language makes it seem like it was an expectation.  I
don't feel that is in the spirit of what the TC has been advocating so
far.

Change-Id: Ifd9d48356ebff23556f24a57224936cf670daba2