Added supplemental files for testing 2015.04
Added a directory of supplemental files for helping to run
Defcore for the 2015.04 standard. Includes a list of all api
tests for the suggested Tempest version, a list of the required
Defcore tests, the simple python tool to generate the required
test list, and a prodecure document for running tests.

Change-Id: Icd93fe30b90cf29fd52b2da3198f2bd6e0362ab0