Flag test_get_console_output
In the compute-servers sections of tests, the
ServerActionsTestJSON.test_get_console_output test fails on some Nova
backends because it's not a universally supported feature.  For example,
it isn't supported on VMware:

http://git.io/vJIAj

This would cause any cloud/product using such a backend to fail DefCore
tests, so the test should be flagged.

Change-Id: Ie3778cf8d1100fdb0a9065fabfb8254e0a7747b1