Flag compute stop/start and lock/unlock tests
In Nova, the various server actions can be limited by the
policy.json file deployed with the API. This policy
file provides the vendor with granular control over the actions
that can be performed by users in different roles. This means
that the “admin-ness” of a feature can be determined by
configuration. For example, in Rackspace’s OpenStack
public cloud the stop/start and lock/unlock behaviors are
not available to standard users and reserved for
administrators only. Because of this, it would be impossible
for these tests to run successfully in the scope of a
public user.

These two pairs of operations are only an example. It is
likely that other vendors would have similar constraints,
so for that reason I’m proposing that these tests be flagged.

Change-Id: I26e012b138e4c78d25f1dd57a4f845e82dc1fdb8