Flag test_list_servers_negative due to Tempest bug
This test currently has a bug that occurs when the Tempest
is configured to not use its isolated tenant feature.
This feature creates new accounts on the fly for individual
tests. This feature can be disabled if the test environment
does not allow for generation of accounts on the fly. This
bug emerges when the Tempest configuration is set to not
allow tenant isolation, but the test overrides that
configuration. From the comments in the Tempest code,
the expected behavior in this situation is that the
test should be skipped
(https://github.com/openstack/tempest/blob/master/tempest/common/credentials.py#L31).
Instead, an exception is raised, which causes the test to
erroneously fail.

I have filled issue with Tempest concerning this issue
(https://bugs.launchpad.net/tempest/+bug/1452118).
I propose that this test be flagged until this bug is resolved.

Change-Id: Ic0f043b8248620eade9af1f9a9033a027a0cbaec