Flagging Compute tests that rely on the Glance V1
In the Compute tests, there are tests that need to generate
images as a prerequisite before the main tests can begin.
Since these are Compute tests, I would expect that these
prerequisites would be via the Compute API, but they use
the Glance v1 API instead. Since the capability being tested
should be Compute in isolation, I would say this test should
be flagged until the depdency on Glance v1 is removed.

Change-Id: I122034315463589ab81d5c549da6142e71c91a0e