Flag compute resize tests
In the compute-servers sections of tests, there are four resize tests.
When using Nova's libvirt driver, and no shared-storage, resize relies
on transferring storage via ssh.

This requires password-less ssh to be setup between all compute nodes
in the cloud. This kind of setup may raise security concerns.

For example, resize is disabled by default in the HP Helion OpenStack
distribution because of this security concerns, as well as in the HP
Helion OpenStack public cloud.

This would cause any cloud/product sharing the same security concern to
fail or skip these tests, so I'm proposing they should be temporarily
flagged.

The blueprint [0], when implemented, would address the security concerns,
so it would be the first step to remove the flag from resize tests.

[0] https://blueprints.launchpad.net/nova/+spec/migrate-libvirt-volumes

Change-Id: I9273b7edac000007225daa7e0b7f03ef0b6e23ed