Add HACKING file
As we receive more submissions to flag or add tests, it's important
for contributors to know how to submit modifications to the guidelines.
In other OpenStack projects, HACKING files are used to denote "rules of
the road" for contributions.  The rules lists laid out in HACKING files
and the automation around enforcement of the rules ostensibly reduce
patch iterations (and thus reviewer workload) and speed up acceptance
of patches by giving contributors insight into the conventions the project
uses before submitting a patch.

This patch adds a HACKING.rst file for the DefCore repository which
provides rules for incoming changes in order to make contributor's
lives easier.

Change-Id: I21f21385f3e51d228d2a8fbf48eb904b6066c0cf