Request to flag Swift tests related to authentication and access
Concerns have been raised (by both community members and John Dickinson,
the Swift PTL) about a subset of the Swift tests that are currently
included that test ACL policy and authentication to objects in Swift.
John's reccomendation was to flag these tests as they are not testing
the Swift API and to remove them in the future.

Change-Id: I0f543ef8863abc0299545c20c4db7ef7aa930c0b