Guideline for 2015.05 based on 2015.next schema 1.2
The main change for this Guideline is the addition
of the Keystone identity tests.

There is a schema update pending but will not have
sufficient review time before Board approval in May.

Based on community discussion, FLAGGED TESTS were removed
from this guideline to be added back by DefCore in the
next cycle following discussion.

This patch includes updates to the 2015.next file
to reflect approval of 2015.05.

Change-Id: Ia1523cb0ee4ecd6ae951303663e8afbbfbc9a4e2