Revert "Update flagging process to include metadata and bug fixes"
This reverts commit d4e396fa2f68c19624a567db8536a35007513028.

The previous commit adds substantial requirements to the flagging process
and was proposed and merged slightly over 3 hours with only 1 review.  I
have reservations about the additions and strongly feel they need to be
discussed.  Adding them without discussion a few days before we send the
process to the Board for approval is not how we should be making
substantive additions.

Change-Id: If3c928908dd46a73401b21d13d65eaa9ff5d829e