Flagging missed stop test
In a previous pull request, it was agreeded to remove the tests
that needed the "stop" function available to end users.

https://review.openstack.org/#/c/180365/

While performing due dilligence, I realized there was one stop
test that I had missed during the previous commit, so I would
like to flag this test under the same reasoning used in
the previous request.

Change-Id: I2b405052f58be5d7ac48955ced5fcde7ea982e45