Close Scale, Open Flag in ProcessCycles.rst
Last week at the OpenStack Liberty Design Summit in Vancouver we held
a DefCore workshop that served as the first meeting of the "Flag"
cycle.  During the meeting we discussed a few objectives and started
our first Flag cycle etherpad [1].  Thus, it's time to tidy up some of
our process documentation.

This patch adds a section for the Flag cycle and some objectives for it
that were discussed at prior meetings in the Scale cycle and at the
inaugural Flag cycle meeting.  It also adds all missing meeting
information for the Scale cycle to the appropriate section, thus paving
the way for us to eventually remove it from the wiki (which will become
very cluttered if we don't prune it once in a while) but still have it
archived.

[1] https://etherpad.openstack.org/p/DefCoreFlag.1

Change-Id: I272c07eb232537f0981f187a3ec74e37ea005a72