Add test_list_get_volume_attachments under compute-volume
test_list_get_volume_attachments is a good replacement for the test:

 tempest.api.compute.volumes.test_attach_volume.AttachVolumeTestJSON.test_attach_detach_volume

test_attach_detach_volume was flagged in I9e79090f70648661ab236f67be7bbcd99bc0afc3
because it requires SSH and is currently not run in the gate.

test_list_get_volume_attachments on the other hand is run in the gate
and is a very basic test of attaching a volume.

NOTE: Since ironic does not support attaching a volume, approving this
patch means any ironic backed cloud will not pass this test. Adding this
test in though because saying you support volumes without testing the
ability to use them (attach/detach) is just silly.

Co-Authored-By: Joe Gordon <joe.gordon0@gmail.com>

Change-Id: Ibedae6f0a89a685da1b59cd1ef1fdbd2a5dbb539