Drop flag on tests that require SSH
Original commit message:
"Currently tests that require SSH are not gated against in Tempest
due to the variety of conflicting network models available to
OpenStack clusters. Tests that require SSH need to be rationalized
and refactored, and is a current work item for the openstack-qa
team.

Until this test infrastructure is refactored, the tests should be
flagged but kept for future capability testing."

Tempest has a blueprint [1] out for a new SSH auth strategy, which
is dependent on several patches landing.  All of the patches
which are necessary for the tests that are covered by the flags
described above have now landed.[2][3]  Therefore, we no longer
need to carry forward the flags on these tests that were present
in prior releases.  Additionally, a patch [4] has been merged in
refstack-client to update the default version of Tempest to a
version new enough to pick up the fixes to Tempest.

[1] https://blueprints.launchpad.net/tempest/+spec/ssh-auth-strategy
[2] https://review.openstack.org/#/c/165504/
[3] https://review.openstack.org/#/c/151601/
[4] https://review.openstack.org/203077/

Co-Authored-By: Chris Hoge <chris@openstack.org>
Co-Authored-By: Mark T. Voelker <mvoelker@vmware.com>

Change-Id: I7d9e273a4324500c9aaae42de942d3b45ffce824