Re-introduce test removal clause
In fd6aca6 [1] we changed rule D302 in HACKING so that tests which
never met DefCore criteria in the first place could never be removed
from the .next.json file (meaning that the flag would stay in place
indefinitely and the test would be re-evaluated at the start of
every DefCore cycle forever).  Per discussion at the Flag.3 meeting [2],
it seems that many Committee members feel that this was an oversite
during review and that it would preferable for such tests to be removed
in the interest of reducing clutter and avoiding spending a lot of
cycles re-evaluating an ever-growing list of tests that are unlikely to
ever meet DefCore criteria.  This patch re-inroduces the clause allowing
such tests to be removed.

 [1] https://review.openstack.org/#/c/185158/
 [2]
http://eavesdrop.openstack.org/meetings/defcore/2015/defcore.2015-06-11-01.02.html

Change-Id: Id63956e9679562ad2c2ee5ec6b6d49cc80dfa6f0