Updated 2015.next.json to use new capabilities
The capabilities classifications were updated to have a more
defined breakdown of the capabilities. This patch takes all of
the existing tests in the 2015.05.next file and reclassifies
them. No tests were added or removed. Patch includes scripts
and supporting files for review, updating, and collaboration.

Also added auth-token as required for compute.

Please review if auth-token should be required for object.

Remove temp files before merging.

Work in progress

Change-Id: Ie3a77c443f30327606ba62832cb6bea2d3ada14e