Flag tests that depend upon SSH in 2015.05.
Original commit message:
"Currently tests that require SSH are not gated against in Tempest
due to the variety of conflicting network models available to
OpenStack clusters. Tests that require SSH need to be rationalized
and refactored, and is a current work item for the openstack-qa
team.
Until this test infrastructure is refactored, the tests should be
flagged but kept for future capability testing."

This situation hasn't changed in the releases covered by 2015.05
and therefore warrants flagging in that Guideline as well.  These
tests are also under consideration for flagging in 2015.json.next [1].

[1] https://review.openstack.org/#/c/189927/1

Change-Id: If536e1afe679f5be400459145ea2be75771fac96