Flag Compute tests that rely on Glancev1 in 2015.05
Original flag commit message:
"In the Compute tests, there are tests that need to generate
images as a prerequisite before the main tests can begin.
Since these are Compute tests, I would expect that these
prerequisites would be via the Compute API, but they use
the Glance v1 API instead. Since the capability being tested
should be Compute in isolation, I would say this test should
be flagged until the depdency on Glance v1 is removed."
Glance implements two APIs, v1 and v2. The v1 API
is currently SUPPORTED, while the v2 API is CURRENT [1][2].
The original plan was to deprecate and remove the v1 API
by Kilo, but it appears it was superceded [3]. Neither Nova
nor Horizon currently supports proxying the v2 API,
but this functionality is on the roadmap for both
projects [4][5].
These tests should be flagged with the intention to
reintroduce them after upstream support for Glance v2 has
been added to Nova and Horizon and after the implicit
dependency on the v1 API has been removed from the Tempest
tests [6].
[1] http://developer.openstack.org/api-ref-image-v1.html
[2] http://developer.openstack.org/api-ref-image-v2.html
[3]
http://lists.openstack.org/pipermail/openstack-dev/2013-November/019431.html
[4] https://blueprints.launchpad.net/nova/+spec/use-glance-v2-api
[5] https://blueprints.launchpad.net/horizon/+spec/horizon-glance-v2
[6] https://bugs.launchpad.net/tempest/+bug/1463534
"

This situation hasn't changed in the releases covered by 2015.05 and
therefore these flags should be carried forward to that Guideline.

Change-Id: I2d8168d1f0478499be574e695bb00ac8e7965c37