Flag test that makes assumption about image creation
...in 2015.05.  Original commit message [1]:
"The Images v2 API ListImages test creates images to appear in the image
list, using the Glance image-create call to do this.  There is, however,
an alternative route to image creation in the Images v2 API, namely, to
use an import task.  A cloud provider may wish to restrict the standard
image-create to admin use only, for example, and restrict regular end
users to use of the import task (which allows an opportunity to perform
validation or security scanning on the image data before creating an
image in Glance).  It is also reasonable for a provider not to allow
any upload of external images into a cloud and instead only allow image
creation via the Nova image-create server action.  The key point is
that given that the ListImages test is supposed to test that the image
list functionality is present and working correctly, the test should be
agnostic as to how the images got into Glance in the first place.  The
test should be flagged until it can be modified to take into account the
various legitimate restrictions that providers may wish to place on how
images are introduced into their OpenStack clouds.

This is also reported as tempest bug:

  https://bugs.launchpad.net/tempest/+bug/1453265 "

The Tempest folks have rejected the bug, but a larger discussion on
the defcore-committee ML [2] has ultimately indicated that there's simply
not an interoperable solution today.  The discussion will hopefully
lead to a more interoperable future, but we're not there yet.  Thus,
we should carry this flag forward to 2015.05.

[1] https://review.openstack.org/#/c/181280/
[2]
http://lists.openstack.org/pipermail/defcore-committee/2015-June/000823.html

Change-Id: If2c6528af24f19881dd30106437bd4253dcc1a1a