Flag Swift ACL test on adivce from Swift PTL
Original commit message [1]:

"Concerns have been raised (by both community members and John
Dickinson, the Swift PTL) about a subset of the Swift tests that
are currently included that test ACL policy and authentication to
objects in Swift.  John's reccomendation was to flag these tests
as they are not testing the Swift API and to remove them in the
future."

Note that these tests have also been proposed for removal
in 2015.next.json. [2]

[1] https://review.openstack.org/#/c/181659/
[2] https://review.openstack.org/#/c/189979/

Change-Id: I445e9b3d8001e286b7147b6ac45beb953c23549e