Move Guidelines into doc directory
As we add more and more Guidelines over time, the top-level
directory of the defcore repo is becoming increasingly cluttered.
Further, each Guideline includes an RST file that could be
rendered on docs.openstack.org and/or gate-checked for validity
like other RST documents.  This patch moves the Guideline docs
into a dedicated folder in the doc directory using a structure
akin to that of the process docs that already reside there. [1][2]
It also cleans up a few references to documents already in the
doc directory that were missed earlier, and adds a note to the
README indicating what the doc/sources/guidelines directory
contains.

Note that a few links on http://openstack.org/interop and DefCore
wiki pages may need to be updated if/when this change lands.  It
is therefore suggested that this patch be merged only when the
relevant Foundation staff are prepared to make corresponding
updates to the website in order to avoid broken links.

[1] https://review.openstack.org/#/c/189950/
[2] https://review.openstack.org/#/c/189956/

Change-Id: Iddd8a4c191c490f4b0afb8c88c8520da4b97576b