Update RST document formatting
The .rst Guideline files currently say that they're generated from
"the master JSON verison" but don't actually contain a link to their
JSON counterparts.  This patch makes that reference a hyperlink so
folks can more easily locate the JSON files that are the authoritative
documents.  This is especially useful when viewing rendered versions
of the RST docs.

The patch also adjusts the script we use to generate the RST docs
from their JSON sources so future Guidelines include similar
linkage.

While we're at it, we've also removed a chunk of code in the
RST generator script so that it doesn't rely on the "name" field
since we never really added those fields universally, which could
cause the script to fail.  Instead we just use the "project" field.

Further, a few capabilities that are marked as removed are lacking
the aforementioned "project" field, so those have been added where
they were missing.

Following the changes to the RST generating script, it has been
re-run so that the existing 2015 Guideline RST documents have
the same formatting.  While the formatting has changed (added links
and modified capitalization), the content has not (no capabilities
or tests added/removed).

Co-Authored-By: Mark T. Voelker <mvoelker@vmware.com>
Co-Authored-By: Rob Hirschfeld <rob@zehicle.com>

Change-Id: Idbdd7d6825aa8af2404b9e2791fa67240cdef239