Add superceded status for Guidelines and 2015B process
NOTE: This patch starts the process of creating a new
draft process for Board approval in Section C6.

As discussed in the board meeting, this change specifically
identifies that only the latest two guidelines are available
for vendor validation.  Older guidelines are still approved
and must remain available because vendors have validated
against them (thus superseded not deprecated).

Change-Id: Id2c2a23c9a8e204480dcebde6fe436292120ba8a