Fix Generator To Reflect New RST Locations
We recently moved the RST guidelines into the doc tree so they
can be eventually be published.  Due to the timing of the change,
we missed the 2015.07 Guideline, which was just approved.  We also
need to update the jsonToRst.py script to reflect the new location.
This patch catches us up on both fronts.

Change-Id: Ia041309b9738ffc8007a054ba611dd6d66943961