Add scoring sheet
At the DefCore sprint in Austin last week we agreed on a format
for moving scoring activities into Gerrit.  This should make
discussion outside of meetings easier to accomplish and make
use of the same toolchain we use for other activities (as well
as having the benefit of being accessible to folks in countries
that cannot access Google Docs where working materials were
formerly kept).

This patch creates a directory for working materials and populates
it with an empty scoring sheet.  We will add capabilties and
scores to it in future patches.

Change-Id: I42dfe410edf4412853b9f976eaf017b90c1a8a51