Fix DefCore Docs Gate Jobs
Over the past week or so we've seen docs jobs failing on the DefCore
repo.  The trouble started at around the same time the gate was having
other issues, so we mostly let it slide.  However upon further review,
it appears a change in the run-docs.sh script in infra [1] is actually the
root of the issue.  The problem is that our tox.ini doesn't have
a testenv configuration that most other projects do, and the change
to run-docs.sh exposed that.  This patch should be sufficient to get
our docs builds working again by fixing our tox.ini file.

[1] https://review.openstack.org/#/c/211239/

Change-Id: I771df06d07080daba8ae0d7beb81d7e2a601aa68