Flag change pwd test in 2015.07 & remove in .next
Chagne password is not widely supported across hypervisors
(in fact, it only currently works with Xen [1]) and therefore
fails to meet DefCore criteria.  This test has also been flagged
in previous Guidelines [2] and we've discussed removing it entirely
since it's extremely unlikely that all other backends will ever
implement this capability.

This patch carries forward the flag from previous Guidelines and
removes the test completely from .next.json so we don't have to
re-flag it again in subsequent Guidelines.

[1]
http://docs.openstack.org/developer/nova/support-matrix.html#operation_set_admin_password
[2] https://review.openstack.org/#/c/196153/3

Change-Id: I13ed481324cf1b0b62c7c02886bdf02906550cbf