Fix Schema 1.3 violations in 2015.07 and .next
In the 2015.07 Guideline, we switched to a new schema [1].  However,
as published today, the 2015.07 Guideline contains a couple of
schema violations.  Namely, it uses strings where it should use
booleans for the "admin" key [2] and it lacks the "status" key for
individual capabilities [3].  This patch corrects both problems, and
also applies them to 2015.next so we don't fall into the same trap
again next time around.

[1]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/schema/1.3.rst
[2]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/schema/1.3.rst#n65
[3]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/schema/1.3.rst#n66

Change-Id: I66949a28efce6f23e6c3371c8ef8356866752beb