remove "functional information" subheading from main toctree
With heading level of the "Functional Information" subsections of the
two 2015 process pages the same as the title of those pages the
subsections appear in the main table of contents, even though they do
not represent a standalone document. This change lowers the heading
level to make it clear the subsections are not meant to appear in the
main table of contents.

Change-Id: Ia7d9c081feedeb556a3a94e7df70b64d310eee32