Bump year of .next.json doc, correct releases
Per our current operating procedures[1], the proposed next Guideline
should have the name "year.next".  The next Guideline will be the first
to go up to the Board of Directors for approval in 2016, so it should
be named "2016.next".  Currently it's named 2015.next, so this patch
corrects the year.  It also updates a couple of scripts that use
the file to point to the right place.  Pointing to the correct year
is not only the procedurally correct thing to do, but it should help
folks understand a little more obviously when the next Guideline would
take effect (something that has come up in a couple of reviews lately).
On a similar vein, this patch also updates the .next.json file to
contain the correct releases covered (bumping it from Icehouse, Juno,
and Kilo to Juno, Kilo, and Liberty).  Here again, this should help
folks understand what releases are covered by the next Guideline.

[1]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/2015A.rst#n61

Change-Id: Ieb8d3f51647c2b18e59a8dc37c1385ba29bc5671