Get rid of year in .next
Currently the filename of the next proposed Guideline is
"[year].next.json" where [year] is the year in which the Board of
Directors will be asked to approve the Guideline.  This is somewhat
problematic in that we now have scripts that refer to this document,
and it's name changes every year (meaning we also have to remember
which scripts are using the old name and update them, as in [1]).

This patch proposes a change to our procedures such that the next
Guideline's filename is static.

[1] https://review.openstack.org/223234

Change-Id: I7af0988f625cbfc94807708f4db2e7262521a606