Define Schema 1.4
As we've begun using Schema 1.3, we've found a couple of minor issues
with it.  First, 1.3 has a duplicate source of truth for the status
(required, advisory, deprecated, removed) of a Capability: there's
a status field in the Capability definition itself, and there are lists
in the Components block of which Capabilities are required, advisory,
deprecated, and removed.  This patch removes the status field from
the Capability block as RefStack currently ignores it anyway [1] (making
this the least-breaking option of the two).

Second, we've identified the need for the next proposed Guideline to
have a static filename, but also let people know when it's being
targetted for a vote by the Board of Directors.[2][3]  I've added a new
field to provide that.

[1] See Paul Van Eck's comment in
https://review.openstack.org/#/c/217393/

[2] See Chris Hoge and Mark Voelker's comments in:
https://review.openstack.org/#/c/223234/

[3] https://review.openstack.org/#/c/223243/

Change-Id: I6102f2b5b0e6eec7ce2ca36249cf0e3bd281d7ef