Add field to record cutoff score
The DefCore process documents spell out that the DefCore Commitee
selects a minimum score that a Capbility must achieve in order to
be required in a Guideline (refer to section A4(4) of the 2015A
processes document).  That minimum score could change from Guideline
to Guideline, as it's part of the DefCore Committee's job to pick a
number with each cycle.  In the past, we've never had a place where
the minimum score was recorded.  As this is information that's
fundamental to Capability inclusion, it seems reasonable that this
would be thing we want to record.  As it may change from cycle to
cycle, it seems reasonable to put it into the schema so it can
be recorded on a per-Guideline basis.

This patch updates Schema 1.4 to add such a field (and uses an
example value that's indicative of the score we've used in the past
as a cutoff).

Change-Id: I2bfc084692963c32285596be3e984eefdea1b1f6