Add aliases field to schema 1.4
The name of a Tempest test that is required by recent DefCore Guidelines
was recently changed [1].  DefCore had not previously considered a way
to handle this scenario: the test list in the Guidelines is considered
immutable.  Even if we changed the name, folks running a version of
Tempest prior to the rename would not get a compliant result.  If did
nothing, folks running a version of Tempest that includes the rename
would not get a compliant result.  We also considered flagging the
tests [2] to alleviate the problem, but there was a general feeling that
since the test hand't actually changed we were reluctant to make the
underlying capability not required.

After some discussion [3] we hit upon the idea of adding an "aliases"
field to our JSON documents.  This would allow us to specify multiple
names for the same test.  Refstack could then read the field and accept
a passing result for either test name when displaying results.  This
patch adds the new alias field to the test block.

[1] https://review.openstack.org/#/c/223796/
[2] https://review.openstack.org/#/c/229177/
[3]
http://eavesdrop.openstack.org/meetings/defcore/2015/defcore.2015-10-07-15.00.log.html#l-23

Change-Id: Iff99fbb50c9b3e5101503de0b032b12c7451aecd