Test names used in the Aliases element.
For future guideline, the test name should be the new name and the
alias should be the old name. This will help RefStack's checking to
be more efficient. When RefStack checks against a spec, the checking
will start from the test name, if there is a match with the result
test list, it does not need to check the Aliases element.  For
prior approved guideline, the Aliases element should hold the new
test names for consistency with the approved doc.

Change-Id: Id87e121f825660c5a0df85f4842271741f928df8