Add the procedure.rst for defcore version 2015.07
In the defcore version 2015.05, if only run tempest, the result is not json
file, so we have no way to use it.

In the defcore version 2015.07, we recommend run testcase of tempest under
refstack-client. When the test is finished, there are two copy files in the
<.tempest/.testrepository/>. The one is xxx.json, that is only pass testcases
in the file. The other file is xxx about detail infos.

In this patch, add a method about how to upload the result(json file) to
refstack-server and check it.

Change-Id: I61ed39730149b6f038d8e0b3793fed6e35219333