Migrate next and 2016.01 to Schema 1.4
At the DefCore workshop at the OpenStack Summit in Tokyo, we
discussed the need to migrate our next.json and in-review 2016.01
Guidelines to the new schema 1.4.  The new schema removes a duplicate
source of truth around the status of each capability, provides a
target date, and an alias field for tests that have been renamed.

This patch migrates the two as-yet unapproved JSON docs to the
new schema.  We've also discussed migrating older specs, which will
be done in a subsequent patch.

Change-Id: I7f0a370d4fd3e2650d70a29accdd17f4894e5a8f