Migrate 2015.07 to Schema 1.4
At the DefCore workshop at the OpenStack Summit in Tokyo, we
discussed the need to migrate our older specs to a newer version
of the schema.  Doing so would make it easier to deal with tests
that have been renamed, and would mean fewer schemas for RefStack
to have to parse.

This patch migrates 2015.07 from schema 1.3 to 1.4.  Older Guidelines
are currently in schemas 1.2 or older and are therefore a bit more
complex to migrate--so those will be dealt with in a separate patch.

Change-Id: I2090166eefeb9157636ecb34a78b061422662c60