Remove volumes-v2-copy-volume-to-image (no tests)
volumes-v2-copy-volume-to-image was included in an earlier change.
During the review [1] it was noted a couple of times that tests were
missing for this capability, but the patch still landed with
the capability included.  DefCore doesn't consider Capabilities that
have no tests [2] since that leaves us without a way for vendors to
demonstrate compliance with an established definition of "working", so
this needs to be removed.

[1] https://review.openstack.org/#/c/221631/
[2]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/CoreDefinition.rst#n131

Change-Id: I943a74181d0aed2dc1f607ededa7df3d3d485c03