Flag tests requiring multiple tenants and users
Guiding core priciple number 7 from the DefCore core
definition is [1]:

"7. Tests can be remotely or self-administered"

Tests that require multiple tenants and users
create a barrier to remote adminstration and
independent verification of test results. Requiring
users to enroll for multiple tenant-isolated accounts
on a public cloud creates a resource cost that
may be insurmountable for end users, especially
when comparing multiple public clouds.

In addition, tenant isolation is less a capability
for writing interoperable applications, and more
a security concern for cloud implementers. While
important, the goal of interopearbility is
orthgonal to the security goals of tenant isolation.

The flagged tests in this patch require multiple
tenant-isloated accounts, and are marked to
to be fixed where possible, or removed where not
possible.

This list was determined by successfully running
the defcore suite against a test cloud[2], then
re-running the tests configured with only one
account against the same test cloud[3].

At the January QA meetup, this test list was
refined by refactoring Server Negative tests
into two classes, one that requires multi-tenant
access, one one that does not[4]. This significantly
reduced the number of flagged tests.[5]

[1] https://github.com/openstack/defcore/blob/master/doc/source/process/CoreDefinition.rst
[2] https://refstack.openstack.org/#/results/0b8ac51d-3288-42d8-b276-4a54046e41bd
[3] https://refstack.openstack.org/#/results/01f29358-8870-4e15-8f7a-f6a082d44e29
[4] https://review.openstack.org/#/c/283819/
[5] https://etherpad.openstack.org/p/multi-tenant-defcore-tests

Change-Id: I5ec801ae48e8220abdaecebe43e77465684e8b5b